requirements.txt
streamlit>=1.30.0
openai>=1.10.0
reportlab>=4.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
